+++ rest-api/alembic/versions/4aa79d5ac86e_add_biobank_history.py
@@ -1,0 +1,165 @@
+"""add biobank history
+
+Revision ID: 4aa79d5ac86e
+Revises: e44e303ae759
+Create Date: 2018-08-20 10:33:58.770173
+
+"""
+from alembic import op
+import sqlalchemy as sa
+import model.utils
+
+
+from participant_enums import PhysicalMeasurementsStatus, QuestionnaireStatus, OrderStatus
+from participant_enums import WithdrawalStatus, SuspensionStatus, QuestionnaireDefinitionStatus
+from participant_enums import EnrollmentStatus, Race, SampleStatus, OrganizationType, BiobankOrderStatus
+from participant_enums import MetricSetType, MetricsKey
+from model.site_enums import SiteStatus, EnrollingStatus, DigitalSchedulingStatus, ObsoleteStatus
+from model.code import CodeType
+
+# revision identifiers, used by Alembic.
+revision = '4aa79d5ac86e'
+down_revision = 'e44e303ae759'
+branch_labels = None
+depends_on = None
+
+
+_BACKFILL_VERSION = """
+    update biobank_order
+    set version = 1
+    """
+_BACKFILL_HISTORY = """
+    insert into biobank_history (biobank_order_id,
+                                 participant_id,
+                                 log_position_id,
+                                 created,
+                                 collected_note,
+                                 processed_note,
+                                 finalized_note,
+                                 finalized_site_id,
+                                 finalized_username,
+                                 source_site_id,
+                                 collected_site_id,
+                                 collected_username,
+                                 processed_site_id,
+                                 processed_username,
+                                 source_username,
+                                 amended_biobank_order_id,
+                                 amended_reason,
+                                 amended_site_id,
+                                 amended_time,
+                                 amended_username,
+                                 cancelled_site_id,
+                                 cancelled_time,
+                                 cancelled_username,
+                                 last_modified,
+                                 order_status,
+                                 restored_site_id,
+                                 restored_time,
+                                 restored_username,
+                                 version
+                                 ) select biobank_order_id,
+    participant_id,
+    log_position_id,
+    created,
+    collected_note,
+    processed_note,
+    finalized_note,
+    finalized_site_id,
+    finalized_username,
+    source_site_id,
+    collected_site_id,
+    collected_username,
+    processed_site_id,
+    processed_username,
+    source_username,
+    amended_biobank_order_id,
+    amended_reason,
+    amended_site_id,
+    amended_time,
+    amended_username,
+    cancelled_site_id,
+    cancelled_time,
+    cancelled_username,
+    last_modified,
+    order_status,
+    restored_site_id,
+    restored_time,
+    restored_username,
+    version  from biobank_order
+    """
+
+
+def upgrade(engine_name):
+    globals()["upgrade_%s" % engine_name]()
+
+
+def downgrade(engine_name):
+    globals()["downgrade_%s" % engine_name]()
+
+
+def upgrade_rdr():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('biobank_history',
+    sa.Column('biobank_order_id', sa.String(length=80), nullable=False),
+    sa.Column('source_username', sa.String(length=255), nullable=True),
+    sa.Column('collected_username', sa.String(length=255), nullable=True),
+    sa.Column('processed_username', sa.String(length=255), nullable=True),
+    sa.Column('finalized_username', sa.String(length=255), nullable=True),
+    sa.Column('order_status', model.utils.Enum(BiobankOrderStatus), nullable=True),
+    sa.Column('amended_reason', sa.UnicodeText(), nullable=True),
+    sa.Column('last_modified', model.utils.UTCDateTime(), nullable=True),
+    sa.Column('restored_username', sa.String(length=255), nullable=True),
+    sa.Column('restored_time', model.utils.UTCDateTime(), nullable=True),
+    sa.Column('amended_username', sa.String(length=255), nullable=True),
+    sa.Column('amended_time', model.utils.UTCDateTime(), nullable=True),
+    sa.Column('cancelled_username', sa.String(length=255), nullable=True),
+    sa.Column('cancelled_time', model.utils.UTCDateTime(), nullable=True),
+    sa.Column('created', model.utils.UTCDateTime(), nullable=False),
+    sa.Column('collected_note', sa.UnicodeText(), nullable=True),
+    sa.Column('processed_note', sa.UnicodeText(), nullable=True),
+    sa.Column('finalized_note', sa.UnicodeText(), nullable=True),
+    sa.Column('version', sa.Integer(), nullable=False),
+    sa.Column('cancelled_site_id', sa.Integer(), nullable=True),
+    sa.Column('finalized_site_id', sa.Integer(), nullable=True),
+    sa.Column('amended_biobank_order_id', sa.String(length=80), nullable=True),
+    sa.Column('restored_site_id', sa.Integer(), nullable=True),
+    sa.Column('processed_site_id', sa.Integer(), nullable=True),
+    sa.Column('collected_site_id', sa.Integer(), nullable=True),
+    sa.Column('participant_id', sa.Integer(), nullable=False),
+    sa.Column('source_site_id', sa.Integer(), nullable=True),
+    sa.Column('log_position_id', sa.Integer(), nullable=False),
+    sa.Column('amended_site_id', sa.Integer(), nullable=True),
+    sa.ForeignKeyConstraint(['amended_biobank_order_id'], ['biobank_order.biobank_order_id'], ),
+    sa.ForeignKeyConstraint(['amended_site_id'], ['site.site_id'], ),
+    sa.ForeignKeyConstraint(['cancelled_site_id'], ['site.site_id'], ),
+    sa.ForeignKeyConstraint(['collected_site_id'], ['site.site_id'], ),
+    sa.ForeignKeyConstraint(['finalized_site_id'], ['site.site_id'], ),
+    sa.ForeignKeyConstraint(['log_position_id'], ['log_position.log_position_id'], ),
+    sa.ForeignKeyConstraint(['participant_id'], ['participant.participant_id'], ),
+    sa.ForeignKeyConstraint(['processed_site_id'], ['site.site_id'], ),
+    sa.ForeignKeyConstraint(['restored_site_id'], ['site.site_id'], ),
+    sa.ForeignKeyConstraint(['source_site_id'], ['site.site_id'], ),
+    sa.PrimaryKeyConstraint('biobank_order_id', 'version')
+    )
+      # ### end Alembic commands ###
+    op.execute(_BACKFILL_VERSION)
+    op.execute(_BACKFILL_HISTORY)
+
+def downgrade_rdr():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('biobank_history')
+    # ### end Alembic commands ###
+
+
+def upgrade_metrics():
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade_metrics():
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
